+++ .github/workflows/pr-build.yaml
@@ -416,8 +416,6 @@ jobs:
         uses: actions/checkout@v2
         with:
           ref: ${{ github.event.pull_request.head.sha }}
-      - name: Log into registry
-        run: echo "${{ secrets.GITHUB_TOKEN }}" | docker login ghcr.io -u ${{ github.actor }} --password-stdin
       - name: Setup environment
         env:
           PR_NUMBER: ${{ github.event.pull_request.number }}
