+++ hack/ci/setup-env.sh
@@ -24,20 +24,18 @@ CERT_NUMBER_TO_BACKUP=1
 CERT_SERVICE_NAMESPACE=capact-system
 EOT
 
-
-
 if [ "${GITHUB_EVENT_NAME}" = "pull_request_target" ]
 then
   echo "DOCKER_TAG=PR-${PR_NUMBER}" >> "$GITHUB_ENV"
-  echo "DOCKER_REPOSITORY=ghcr.io/mkuziemko/pr" >> "$GITHUB_ENV"
+  echo "DOCKER_REPOSITORY=ghcr.io/capactio/pr" >> "$GITHUB_ENV"
 else
   echo "DOCKER_TAG=${GITHUB_SHA:0:7}" >> "$GITHUB_ENV"
-  echo "DOCKER_REPOSITORY=ghcr.io/mkuziemko" >> "$GITHUB_ENV"
+  echo "DOCKER_REPOSITORY=ghcr.io/capactio" >> "$GITHUB_ENV"
 fi
 
 # TODO: Read components to build in automated way, e.g. from directory structure
 cat <<EOT >>"$GITHUB_ENV"
 APPS=name=matrix::{"include":[{"APP":"gateway"},{"APP":"k8s-engine"},{"APP":"hub-js"},{"APP":"argo-runner"},{"APP":"helm-runner"},{"APP":"populator"},{"APP":"terraform-runner"},{"APP":"argo-actions"},{"APP":"gitlab-api-runner"},{"APP":"secret-storage-backend"},{"APP":"helm-storage-backend"}]}
-TESTS=name=matrix::{"include":[{"TEST":"e2e"}]}
+TESTS=name=matrix::{"include":[{"TEST":"e2e"}, {"TEST":"localhub"}]}
 INFRAS=name=matrix::{"include":[{"INFRA":"json-go-gen"},{"INFRA":"graphql-schema-linter"},{"INFRA":"jinja2"},{"INFRA":"merger"}]}
 EOT
